Fourier_image/requirements.txt
matplotlib==3.5.1
numpy==1.23.3
Pillow==10.0.0
scikit_learn==1.3.0
torch==1.13.1
torchvision==0.12.0+cu113
tqdm==4.63.1
